2017年12月18日

+++ mserver/src/main/java/com/co/example/dao/user/TUserRoleDao.java
@@ -1,7 +1,7 @@
 package com.co.example.dao.user;
 
-import com.github.moncat.common.dao.BaseDao;
 import com.co.example.entity.user.TUserRole;
+import com.github.moncat.common.dao.BaseDao;
 
-public interface TUserRoleDao extends BaseDao<TUserRole, Integer> {
+public interface TUserRoleDao extends BaseDao<TUserRole, Long> {
 }
