+++ mserver/src/main/java/com/co/example/service/user/TUserRoleService.java
@@ -1,7 +1,7 @@
 package com.co.example.service.user;
 
-import com.github.moncat.common.service.BaseService;
 import com.co.example.entity.user.TUserRole;
+import com.github.moncat.common.service.BaseService;
 
-public interface TUserRoleService extends BaseService<TUserRole, Integer> {
+public interface TUserRoleService extends BaseService<TUserRole, Long> {
 }
